K Nearest Neighbors
K nearest neighbors is a non-parametric method used for classification and regression.In both cases, the input consists of the k closest training examples in the feature space. The output depends on whether knn is used for classification or regression. 

In this exercise we are trying to predict the class based on different set of variables.